feat: add region auto-detect
This change will allow substation to auto-detect the region if no
region is defined. If more than one region is returned as part
of the auth catalog an error will be raised instructing the user
to set the region_name value in their configuration file.

Related-Issue: https://github.com/cloudnull/substation/issues/11
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>