chore: reflow / format the docs
This is changing the flow of the docs because it was needed.
I had a lot of information in the docs that was duplicated across
pages, concepts that were througts from old builds, etc. This
change consolodates the docs so things flow a little better
and should be easier to understand.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>