feat: add bulk operations for multi-select
This is the first bulk operation that will allow the application to
multi-delete systems. Now, if a user presses CTRL-X the status screen
will change to a multi-select window, and permit the user to select
many items, which can then be used to initiate a delete. In the future
this will be able to do like bulk snapshots, life-cycle management, and
more. For now we do destructive things because that's easy.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>