feat: implement command style navigation
This change implements command style navigation. Users are now able
to press ":" to bring up a command window that will fuzzy match
on input and update the left side navigation dynamically as the
user types.

There's still a lot to do to make this navigation change complete,
however, this is a big step up in terms of operator usability.

Closes: https://github.com/cloudnull/substation/issues/28
Closes: https://github.com/cloudnull/substation/issues/19
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>